More options to recover from broken usercode
Start the watchdog if START_WATCHDOG is defined (usercode must handle this!), enter bootloader if reboot was caused by a watchdog timeout.
Look for a 'magic' value in the usercode area at PROG_REGION_START + 0x10, don't boot usercode it it's missing. Means we won't boot a not-bootloader-adapted hexfile  anymore.
Check another 'magic' value in RAM, can be used to enter the bootloader from usercode (usercode loads the magic value to a specific address in RAM, then resets. Bootloader will not clear this address, read from it and activate if it contains the magic value).
Start a timeout timer when bootloader was activated with usercode present, so an accidental bootloader activation won't hang the system 'forever'.